requirements.txt
dj-database-url==0.5.0
Django==1.10.5
django-cors-headers==2.4.0
django-heroku==0.3.1
django-webpack-loader==0.5.0
djangorestframework==3.7.3
gunicorn==19.7.1
psycopg2==2.7.5
pytz==2018.5
whitenoise==3.3.1
